#777- text object introduced
